Remove PBS
removes all references to PBS.
This makes way for adding PBS back in, as a plugin!

Refs #9000
