+++ python/TestHarness/testers/RunApp.py
@@ -24,10 +24,6 @@ class RunApp(Tester):
         # it'll be used in lieu of building up the command automatically
         params.addParam('command',            "The command line to execute for this test.")
 
-        params.addParam('walltime',           "The max time as pbs understands it")
-        params.addParam('job_name',           "The test name as pbs understands it")
-        params.addParam('no_copy',            "The tests file as pbs understands it")
-
         # Parallel/Thread testing
         params.addParam('max_parallel', 1000, "Maximum number of MPI processes this test can be run with      (Default: 1000)")
         params.addParam('min_parallel',    1, "Minimum number of MPI processes that this test can be run with (Default: 1)")
@@ -180,90 +176,7 @@ class RunApp(Tester):
         else:
             command = specs['executable'] + timing_string + specs['input_switch'] + ' ' + specs['input'] + ' ' + ' '.join(specs['cli_args'])
 
-        if options.pbs:
-            return self.getPBSCommand(options)
-
         return command
-
-    def getPBSCommand(self, options):
-        if options.parallel == None:
-            default_ncpus = 1
-        else:
-            default_ncpus = options.parallel
-
-        # Raise the floor
-        ncpus = max(default_ncpus, int(self.specs['min_parallel']))
-        # Lower the ceiling
-        ncpus = min(ncpus, int(self.specs['max_parallel']))
-
-        # Set number of threads to be used lower bound
-        nthreads = max(options.nthreads, int(self.specs['min_threads']))
-        # Set number of threads to be used upper bound
-        nthreads = min(nthreads, int(self.specs['max_threads']))
-
-        extra_args = ''
-        if options.parallel or ncpus > 1 or nthreads > 1:
-            extra_args = ' --n-threads=' + str(nthreads) + ' ' + ' '.join(self.specs['cli_args'])
-
-        timing_string = ' '
-        if options.timing:
-            self.specs['cli_args'].append('--timing')
-            self.specs['cli_args'].append('Outputs/print_perf_log=true')
-
-        # Append any extra args to the cluster_launcher
-        if extra_args != '':
-            self.specs['cli_args'] = extra_args
-        else:
-            self.specs['cli_args'] = ' '.join(self.specs['cli_args'])
-        self.specs['cli_args'] = "'" + self.specs['cli_args'].strip() + "'"
-
-        # Open our template. This should probably be done at the same time as cluster_handle.
-        template_script = open(os.path.join(self.specs['moose_dir'], 'python', 'TestHarness', 'pbs_template.i'), 'r')
-        content = template_script.read()
-        template_script.close()
-
-        # Convert MAX_TIME to hours:minutes for walltime use
-        hours = int(int(self.specs['max_time']) / 3600)
-        minutes = int(int(self.specs['max_time']) / 60) % 60
-        self.specs['walltime'] = '{0:02d}'.format(hours) + ':' + '{0:02d}'.format(minutes) + ':00'
-
-        # Truncate JOB_NAME. PBS can only accept 13 character (6 characters from test name + _## (test serial number) + _### (serialized number generated by cluster_launcher) = the 13 character limit)
-        self.specs['job_name'] = self.specs['input'][:6] + '_' + str(options.test_serial_number).zfill(2)
-        self.specs['job_name'] = self.specs['job_name'].replace('.', '')
-        self.specs['job_name'] = self.specs['job_name'].replace('-', '')
-
-        # Convert TEST_NAME to input tests file name (normally just 'tests')
-        self.specs['no_copy'] = options.input_file_name
-
-        # Are we using the PBS Emulator? Make this param valid if so.
-        # Add the substitution string here so it is not visable to the user
-        self.specs.addStringSubParam('pbs_stdout', 'PBS_STDOUT', "Save stdout to this location")
-        self.specs.addStringSubParam('pbs_stderr', 'PBS_STDERR', "Save stderr to this location")
-        if options.PBSEmulator:
-            self.specs['pbs_stdout'] = 'pbs_stdout = PBS_EMULATOR'
-            self.specs['pbs_stderr'] = 'pbs_stderr = PBS_EMULATOR'
-        else:
-            # The PBS Emulator fails when using the PROJECT argument (#PBS -P <project name>)
-            self.specs.addStringSubParam('pbs_project', 'PBS_PROJECT', "Identify this job submission with this project")
-            self.specs['pbs_project'] = 'pbs_project = %s' % (options.pbs_project)
-
-
-        # Do all of the replacements for the valid parameters
-        for spec in self.specs.valid_keys():
-            if spec in self.specs.substitute:
-                self.specs[spec] = self.specs.substitute[spec].replace(spec.upper(), self.specs[spec])
-            content = content.replace('<' + spec.upper() + '>', str(self.specs[spec]))
-
-        # Make sure we strip out any string substitution parameters that were not supplied
-        for spec in self.specs.substitute_keys():
-            if not self.specs.isValid(spec):
-                content = content.replace('<' + spec.upper() + '>', '')
-
-        # Write the cluster_launcher input file
-        options.cluster_handle.write(content + '\n')
-
-        return os.path.join(self.specs['moose_dir'], 'scripts', 'cluster_launcher.py') + ' ' + options.pbs + '.cluster'
-
 
     def processResults(self, moose_dir, retcode, options, output):
         reason = ''
@@ -296,9 +209,6 @@ class RunApp(Tester):
             # Valgrind runs
             elif retcode == 0 and self.shouldExecute() and options.valgrind_mode != '' and 'ERROR SUMMARY: 0 errors' not in output:
                 reason = 'MEMORY ERROR'
-            # PBS runs
-            elif retcode == 0 and options.pbs and 'command not found' in output:
-                reason = 'QSUB NOT FOUND'
 
         # Populate the bucket
         if reason != '':
