+++ python/TestHarness/util.py
@@ -495,7 +495,6 @@ class TestStatus(object):
     ## PASS    =  Passing tests 'OK'
     ## FAIL    =  Failing tests
     ## DIFF    =  Failing tests due to Exodiff, CSVDiff
-    ## PBS     =  Any statuses belonging to messages generated by PBS
     ## PENDING =  A pending status applied by the TestHarness (RUNNING...)
     ## DELETED =  A skipped test hidden from reporting. Under normal circumstances, this sort of test
     ##            is placed in the SILENT bucket. It is only placed in the DELETED bucket (and therfor
@@ -511,7 +510,6 @@ class TestStatus(object):
     bucket_fail    = test_status(status='FAIL', color='RED')
     bucket_deleted = test_status(status='DELETED', color='RED')
     bucket_diff    = test_status(status='DIFF', color='YELLOW')
-    bucket_pbs     = test_status(status='PBS', color='CYAN')
     bucket_pending = test_status(status='PENDING', color='CYAN')
     bucket_skip    = test_status(status='SKIP', color='RESET')
     bucket_silent  = test_status(status='SILENT', color='RESET')
